+++ src/selectors/AppSelectors.js
@@ -1,7 +1,11 @@
 import { createSelector } from 'reselect';
 import { getConfig } from 'config/Config';
+import { createNoteFilterFromDefinition, getDefaultSelectedNoteFilter } from 'data/DataNoteFilters';
+import { createTaskFilterFromDefinition, getDefaultSelectedTaskFilter } from 'data/DataTaskFilters';
+import { getNoteFiltersFilteredByVisibleState } from 'selectors/NoteFilterSelectors';
 import { getSession } from 'selectors/SessionSelectors';
-import { getSettings } from 'selectors/SettingSelectors';
+import { getCategoryNoteSorters, getCategoryTaskSorters, getSettings } from 'selectors/SettingSelectors';
+import { getTaskFiltersFilteredByVisibleState } from 'selectors/TaskFilterSelectors';
 import { verifyLicense } from 'utils/LicenseUtils';
 import { getValue } from 'utils/ObjectUtils';
 
@@ -12,14 +16,14 @@ export const getMinuteTimer = state => state.app.minuteTimer;
 export const getEditingCell = state => state.app.editingCell;
 
 export const getSelectedNoteIds = state => state.app.selectedNoteIds;
-export const getSelectedNoteFilter = state => state.app.selectedNoteFilter;
-export const getSelectedNoteFilterId = state => state.app.selectedNoteFilter ? state.app.selectedNoteFilter.id : null;
+export const getSelectedNoteFilterDefinition = state => state.settings.selectedNoteFilterDefinition;
+export const getSelectedNoteFilterId = state => state.settings.selectedNoteFilterDefinition ? state.settings.selectedNoteFilterDefinition.id : null;
 export const getSelectedNoteFilterDate = state => state.app.selectedNoteFilterDate;
 export const getSearchNoteValue = state => state.app.searchNoteValue;
 
 export const getSelectedTaskIds = state => state.app.selectedTaskIds;
-export const getSelectedTaskFilter = state => state.app.selectedTaskFilter;
-export const getSelectedTaskFilterId = state => state.app.selectedTaskFilter ? state.app.selectedTaskFilter.id : null;
+export const getSelectedTaskFilterDefinition = state => state.settings.selectedTaskFilterDefinition;
+export const getSelectedTaskFilterId = state => state.settings.selectedTaskFilterDefinition ? state.settings.selectedTaskFilterDefinition.id : null;
 export const getSelectedTaskFilterDate = state => state.app.selectedTaskFilterDate;
 export const getSearchTaskValue = state => state.app.searchTaskValue;
 
@@ -80,5 +84,25 @@ export const isPro = createSelector(
     getActivationInfo,
     (activationInfo) => {
         return !!activationInfo;
+    }
+);
+
+export const getSelectedNoteFilter = createSelector(
+    getSelectedNoteFilterDefinition,
+    getNoteFiltersFilteredByVisibleState,
+    getCategoryNoteSorters,
+    (filterDefinition, filters, sorters) => {
+        const filter = createNoteFilterFromDefinition(filterDefinition, filters, sorters);
+        return filter ? filter : getDefaultSelectedNoteFilter();
+    }
+);
+
+export const getSelectedTaskFilter = createSelector(
+    getSelectedTaskFilterDefinition,
+    getTaskFiltersFilteredByVisibleState,
+    getCategoryTaskSorters,
+    (filterDefinition, filters, sorters) => {
+        const filter = createTaskFilterFromDefinition(filterDefinition, filters, sorters);
+        return filter ? filter : getDefaultSelectedTaskFilter();
     }
 );