+++ src/selectors/NoteSelectors.js
@@ -1,8 +1,10 @@
 import moment from 'moment';
 import { createSelector } from 'reselect';
-import { combineConditions, createSearchNoteValueCondition } from 'data/DataNoteFilters';
+import { combineConditions, createNoteFilterFromDefinition, createSearchNoteValueCondition } from 'data/DataNoteFilters';
 import { getSearchNoteValue, getSelectedNoteFilter, getSelectedNoteFilterDate, getSelectedNoteIds } from 'selectors/AppSelectors';
 import { getNoteFieldsIncludingDefaults } from 'selectors/NoteFieldSelectors';
+import { getNoteFiltersFilteredByVisibleState } from 'selectors/NoteFilterSelectors';
+import { getCategoryNoteSorters, getCombinedNoteFilterDefinitions } from 'selectors/SettingSelectors';
 import { isBusy } from 'selectors/ThreadSelectors';
 import { store } from 'store/Store';
 import { filterByVisibleState } from 'utils/CategoryUtils';
@@ -29,8 +31,11 @@ export const getNotesFilteredBySelectedFilter = createSelector(
     getSelectedNoteFilter,
     getSelectedNoteFilterDate,
     getNoteFieldsIncludingDefaults,
+    getNoteFiltersFilteredByVisibleState,
+    getCombinedNoteFilterDefinitions,
+    getCategoryNoteSorters,
     isBusy,
-    (notes, searchNoteValue, selectedNoteFilter, selectedNoteFilterDate, noteFields, busy) => {
+    (notes, searchNoteValue, selectedNoteFilter, selectedNoteFilterDate, noteFields, noteFilters, combinedNoteFilterDefinitions, categoryNoteSorters, busy) => {
         if (busy) {
             return getNotesFilteredBySelectedFilterResult;
         }
@@ -40,6 +45,14 @@ export const getNotesFilteredBySelectedFilter = createSelector(
         if (searchNoteValue) {
             extraConditions.push(createSearchNoteValueCondition(searchNoteValue));
         }
+
+        (combinedNoteFilterDefinitions || []).forEach(filterDefinition => {
+            const filter = createNoteFilterFromDefinition(filterDefinition, noteFilters, categoryNoteSorters);
+
+            if (filter) {
+                extraConditions.push(filter.condition);
+            }
+        });
 
         selectedNoteFilter = combineConditions(selectedNoteFilter, extraConditions);