+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 
 * Show "total until now" in work log tab
 * Move custom filters to the general directory (#100)
+* Restore last used filter on startup (#110)
 * Import/export data (#119)
 * Show next save/backup/synchronization date (#121)
+* Combine filters (#122)
 * Add line between rows in task table (#127)
 * Improve "Add sub-task" action (#134)
 * Synchronize note/task filters and task templates with TaskUnifier Cloud